refactor(core): move FocusedChangedEvent to types
